Add automatic posix shell detection
Many of the tests used by the skip statement in the default mrconfig
assume that the shell evaluating them is POSIX-compliant. The perl
system() and exec() functions call /bin/sh for any subshell.
Unfortunately, /bin/sh on Solaris and other SVR4-based systems is not
POSIX compliant. However /usr/xpg4/bin/sh is. This change adds a
function called get_posix_shell which returns the path to the POSIX
shell on that system, and replaces any system calls that evaluate mr
tests with a call to system($posixshell, '-c', $test)

This was done at the perl level rather than as an mrconfig option
because making $posixshell into a nasty "case" statement seemed too
clunky.